nexus_orchestrator/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0

# AI and Orchestration
crewai==0.28.0
langchain==0.0.319
langchain-openai==0.0.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.5.0
redis==5.0.1
qdrant-client==1.6.4

# API Integration
openai==1.1.1
anthropic==0.5.0
cohere==4.32

# Authentication and Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Web3 and Blockchain
web3==6.10.0
eth-account==0.9.0

# Utilities
aiohttp==3.8.6
tenacity==8.2.3
pydantic-settings==2.0.3
